refactor(hub): extract alert subsystem into alerts.go
Pure code move — no behavior change. Same package, same globals.
hub/main.go: 3768 → 3306 lines, hub/alerts.go: 462 lines.

Moved (still package main):
  - AlertRule, Alert types
  - seedAlertRules, alertEvalLoop, evaluateAlerts, compareValue, getAlertRules
  - broadcastAlertSSE (uses sseClients/sseClientsMu still in main)
  - sendTelegram (uses telegramToken/telegramChatID still in main)
  - handleListAlerts, handleAlertCount, handleAcknowledgeAlert
  - handleListAlertRules, handleUpdateAlertRule

go test -count=1 ./...  green 3 runs in a row.
